+++ src/boosting/mod.rs
@@ -11,6 +11,7 @@ pub(crate) const MAX_BINS: usize = 255;
 pub(crate) const MAX_TREE_NODES: usize = 131_071;
 pub(crate) const MAX_TREE_LEAVES: usize = 65_536;
 pub(crate) const MAX_TREE_DEPTH: usize = 256;
+pub(crate) const MAX_TREES: usize = 4_096;
 
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) enum BoostingError {